+++ src/pages/user/[...displayId].tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
-import UserCenter from '@/components/UserCenter';
+import UserCenter from '@/components/pages/UserCenter';
 import { AuthorProfile } from '@/components/pages/AuthorProfile';
 import { mockAuthor } from '@/components/pages/AuthorProfile/types';
 import { fetchUserProfileServer, ServerUserProfileResponse } from '@/services';
